Add package-lock.json and refactor RealtimeDetail components to use new context
This commit introduces a new package-lock.json file for dependency management. It refactors the RealtimeDetail components to utilize the updated RealtimeDetail context, enhancing the structure and functionality of the application. The changes include replacing the RealtimeDetail and RealtimeDetailHeader components with a new RealtimeCreate structure, improving user experience during the alert creation process. Additionally, the context now handles alert saving and deletion more effectively, ensuring better navigation and error handling.
