+++ alerts/apps/frontend/src/components/realtime/detail/RealtimeDetailSectionValidity/index.tsx
@@ -2,7 +2,7 @@
 
 /* * */
 
-import { useRealtimeDetailContext } from '@/contexts/RealtimeCreate.context';
+import { useRealtimeDetailContext } from '@/contexts/RealtimeDetail.context';
 import { Collapsible, DateTimePicker, Section } from '@tmlmobilidade/ui';
 import { Dates } from '@tmlmobilidade/utils';
