+++ alerts/apps/frontend/src/components/realtime/detail/RealtimeReferencesAgencies/index.tsx
@@ -2,7 +2,7 @@
 
 /* * */
 
-import { useRealtimeDetailContext } from '@/contexts/RealtimeCreate.context';
+import { useRealtimeDetailContext } from '@/contexts/RealtimeDetail.context';
 import { IconPlus, IconTrash } from '@tabler/icons-react';
 import { Button, Combobox, Label, Section, Surface } from '@tmlmobilidade/ui';
 import { useMemo } from 'react';
